pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hra-hackathon"
version = "0.0.1"
description = ""
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "bigraph-schema",
    "process-bigraph",
    "copasi-basico",
    "biomodels",
    "python-libsedml",
    "requests",
    "plotly>=6.5.2",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "build",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["processes*", "experiments*"]
exclude = ["tests*", "docs*"]

[tool.uv.sources]
